Rename data/employmentIncomeDeduction.ts to netEmploymentIncome.ts
The file now models both reductions that turn gross employment income into 給与所得
— the 給与所得控除 tables and the 所得金額調整控除 amount — so name it for that
shared purpose (its core export is already calculateNetEmploymentIncomeForPeriod)
and to read cohesively with NetEmploymentIncomeTooltip. Internal symbol names are
left concept-specific. Updates the 4 importers and the header doc comment.
